+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix the first-start dialog save button ([#423](https://github.com/ben/foundry-ironsworn/pull/423))
 - Move "manual" content from the readme to the [wiki](https://github.com/ben/foundry-ironsworn/wiki/Getting-Started)
+- Run the "Ironsworn Startup" macro on startup, if it exists
 
 ## 1.16.5